06 Simetrično dekriptiranje podataka
- Implementirana funkcija za dekriptiranje kriptiranog sadržaja iz datoteke

+++ Sustav za kriptiranje i dekriptiranje/UpravljanjeDatotekama/Datoteka.cs
@@ -113,7 +113,19 @@ namespace Sustav_za_kriptiranje_i_dekriptiranje.UpravljanjeDatotekama
         public static string KreirajAesKripitraniSadrzajDatoteku()
         {
             string putanjaDatoteke = radniDirektorij + @"/kriptirani_sadrzaj.txt";
-            File.Create(putanjaDatoteke).Close();
+            if (!File.Exists(putanjaDatoteke))
+            {
+                File.Create(putanjaDatoteke).Close();
+            }
+            return putanjaDatoteke;
+        }
+        public static string KreirajAesDekriptiraniSadrzajDatoteku()
+        {
+            string putanjaDatoteke = radniDirektorij + @"/dekriptirani_sadrzaj.txt";
+            if (!File.Exists(putanjaDatoteke))
+            {
+                File.Create(putanjaDatoteke).Close();
+            }
             return putanjaDatoteke;
         }
     }
